requirements.txt
fastapi
uvicorn
mysql-connector-python
pydantic
streamlit
requests
pandas
python-dotenv
